small updates/optimizations: controller\ProfileController -> changePassword & deleteAccount, services\ProfileService -> changePassword & deleteAccount

+++ app/Http/Controllers/ProfileController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\Profile\ChangePasswordRequest;
 use App\Http\Requests\Profile\DeleteAccountRequest;
 use App\Services\ProfileService;
@@ -20,14 +21,22 @@ class ProfileController extends Controller
      */
     public function changePassword(ChangePasswordRequest $request): JsonResponse
     {
+        $user = $request->user();
+
+        // check password
+        if (!Hash::check($request->oldPassword, $user->password)) {
+            return $this->error('Wrong password', 400);
+        };
+
+        // call profile service
         $response = $this->profileService->changePassword(
             $request->user(),
-            $request->old_password,
             $request->new_password
         );
 
+        // return json
         if (!$response) {
-            return $this->error('Wrong password', 400);
+            return $this->error('Failed to change password!', 500);
         }
 
         return $this->success(null, 'Password changed successfully');
@@ -38,14 +47,22 @@ class ProfileController extends Controller
      */
     public function deleteAccount(DeleteAccountRequest $request): JsonResponse
     {
+        $user = $request->user();
+
+        // check password
+        if (!Hash::check($request->password, $user->password)) {
+            return $this->error('Password incorrect!', 400);
+        }
+
+        // call profile service
         $response = $this->profileService->deleteAccount(
             $request,
-            $request->user(),
-            $request->password
+            $user
         );
 
+        // return json
         if (!$response) {
-            return $this->error('Password incorrect!', 400);
+            return $this->error('Failed to delete account!', 500);
         }
 
         return $this->success(null, 'Your account has been deleted');
